Implement getHumanChoice() function
"Step 3" of the project outline.

This implements a function which returns either "rock", "paper", or
"scissors" by calling prompt() and returning the user input. The
instructions for this step say to disregard the possibility of invalid
user input and reprompting, since that will use techniques that are
covered later. I did include a simple while loop that reprompts until
the input matches "rock" "paper" or "scissors" explicitly, but I'll
comment those lines in subsequent commits. For now, these lines are
included.