Implement playRound() function
This is "Step 5" of the project outline.

playRound() takes strings humanChoice and computerChoice and determines
a winning player. It handles the case of ties as well. In all three cases,
it prints a message to console with the outcome of the round, and also
increments the score of the winning player.