Implement playGame() and reportWinner() functions
"Step Six" of the project outline. This is the last step of the project.

Implementing playGame() involves refactoring the previous code, namely by moving
the playRound() function and humanScore and computerScore variables to be nested
within playGame(). After declaring variables for human/computer choices and
declaring and initializing variables for human/computer scores, then declaring
playRound() function, the main loop of playGame() is a for loop that iterates 5
times. With each pass, a new value is assigned to the human/computer choices by
calling getHumanChoice() and getComputerChoice(), and then calling playRound() with
those values as arguments. playRound() has also been refactored to take a third
argument roundNumber, so that it can include the current round with each output to
console.

After the loop terminates, a helper function reportWinner() is called with the
humanScore and computerScore as arguments. A message is output to console, reporting
the score and whether the human player won, lost, or tied.

Finally, implemented in an earlier commit but not documented, there is a helper
function called capitalizeFirstLetter() to allow for capitalizing the humanChoice or
computerChoice as needed in the output to console during playRound().

In the future, an alternative implementation would not allow for ties. Instead,
playRound() could recursively call itself in the event of a tie (after calling
getHumanChoice() and getComputerChoice() again for new values, of course). This
would not require changing the inner loop in playGame(), since it would only be
iterating when playRound() returns.